+++ include/index/ranker/ranker.h
@@ -48,7 +48,7 @@ class ranker
      * @param filter A filtering function to apply to each doc_id; returns true
      * if the document should be included in results
      */
-    std::vector<std::pair<doc_id, double>>
+    std::vector<std::pair<doc_id, float>>
         score(inverted_index& idx, corpus::document& query,
               uint64_t num_results = 10,
               const std::function<bool(doc_id d_id)>& filter = [](doc_id)
@@ -61,14 +61,14 @@ class ranker
      * query term.
      * @param sd The score_data for this query
      */
-    virtual double score_one(const score_data& sd) = 0;
+    virtual float score_one(const score_data& sd) = 0;
 
     /**
      * Computes the constant contribution to the score of a particular
      * document.
      * @param sd The score_data for the query
      */
-    virtual double initial_score(const score_data& sd) const;
+    virtual float initial_score(const score_data& sd) const;
 
     /**
      * Default destructor.
