+++ src/index/ranker/pivoted_length.cpp
@@ -15,18 +15,18 @@ namespace index
 
 const std::string pivoted_length::id = "pivoted-length";
 
-pivoted_length::pivoted_length(double s) : s_{s}
+pivoted_length::pivoted_length(float s) : s_{s}
 {
     /* nothing */
 }
 
-double pivoted_length::score_one(const score_data& sd)
+float pivoted_length::score_one(const score_data& sd)
 {
-    double doc_len = sd.idx.doc_size(sd.d_id);
-    double TF = 1 + fastapprox::fasterlog(
+    float doc_len = sd.idx.doc_size(sd.d_id);
+    float TF = 1 + fastapprox::fasterlog(
                         1 + fastapprox::fasterlog(sd.doc_term_count));
-    double norm = (1 - s_) + s_ * (doc_len / sd.avg_dl);
-    double IDF
+    float norm = (1 - s_) + s_ * (doc_len / sd.avg_dl);
+    float IDF
         = fastapprox::fasterlog((sd.num_docs + 1) / (0.5 + sd.doc_count));
     return TF / norm * sd.query_term_count * IDF;
 }
